+++ upi-qr-code-payment-for-woocommerce.php
@@ -3,14 +3,14 @@
  * Plugin Name: UPI QR Code Payment for WooCommerce
  * Plugin URI: https://wordpress.org/plugins/upi-qr-code-payment-for-woocommerce/
  * Description: It enables a Woocommerce site to accept payments through UPI apps like BHIM, Google Pay, Paytm, PhonePe or any Banking UPI app. Avoid payment gateway charges.
- * Version: 1.0.0
+ * Version: 1.0.1
  * Author: Sayan Datta
- * Author URI: https://about.me/iamsayan
+ * Author URI: https://sayandatta.in
  * License: GPLv3
  * Text Domain: upi-qr-code-payment-for-woocommerce
  * Domain Path: /languages
- * WC requires at least: 3.0
- * WC tested up to: 3.9
+ * WC requires at least: 3.1
+ * WC tested up to: 4.0
  * 
  * UPI QR Code Payment for WooCommerce is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by